feat: create base container for profile screen
