+++ app/src/NativeBase/Containers/index.ts
@@ -6,5 +6,6 @@ export { default as ProfileContainer } from './ProfileContainer'
 export { default as ProjectDetailContainer } from './ProjectDetailContainer'
 export { default as ProjectSearchContainer } from './ProjectSearchContainer'
 export { default as SearchResultsContainer } from './SearchResultsContainer'
+export { default as SettingsContainer } from './SettingsContainer'
 export { default as VerticalStackContainer } from './VerticalStackContainer'
 export { default as WebViewContainer } from './WebViewContainer'
